added a file structure which makes sense
